new main file: just TFLite, not coral
Coral edgeTPU remains out-of-stock. This Main file switches over to using the normal TFLite runtime, and the non-coral model file. Speed is only marginally compromised. Lets use this as the main on-device code to develop on right now.. we can switch to the coral workflow with a few lines of code once the devices are back in stock..